Tweaks currency unit

+++ Sapphire/Domain/Balances/BalanceData.swift
@@ -6,11 +6,11 @@ struct BalanceData {
         let name: String
         let longName: String
         let balance: Double
-        let last: Int64?
-        let high: Int64?
-        let low: Int64?
+        let last: Double?
+        let high: Double?
+        let low: Double?
         let change: Double?
-        let estimatedBTCValue: Int64
+        let estimatedBTCValue: Double
     }
 
     var date: Date
